交差点編集で、トラックの線が表示される。トラックの経路の新規追加分がカーブになる。 (#390)
* 横断歩道を生成

横断歩道を引くべきラインを計算

同上

破線実装中

横断歩道を破線に

交差点を個別ゲームオブジェクトに

横断歩道のゲームオブジェクト名を整理

横断歩道の幅と位置を調整

同上

* 横断歩道の配置方法オプションを追加（追加/編集）

同上

交差点を全配置、配置しないオプション実装

交差点を「大きい道路に配置」オプション

編集の横断歩道配置オプション

道路編集の横断歩道UI

横断歩道を削除できるようにする

* リネーム

* リファクタ、不要なインターフェイスや構造体を削除

不要なインターフェイスを削除中

インターフェイス削除中

同上

不要な箇所を削除

不要な構造体を削除

リファクタ中

リファクタ中

* リファクタ: 利用されていない機能を削除や移動

不要な部分を削除

利用されていない機能を別ファイルに移動

リファクタ

リファクタ

* リネーム： RoadNetworkSceneGUISystem → RoadNetworkEditTargetSelectButton

* fix : GeoGraph2D.RemoveSelfCrossingでcreatorに渡すときにf2が2になっていたので修正. それプラスコメント追加を引数のself -> pointsに変更

* リファクタ: 不要な機能を削除または別ファイルに移動、警告に対処

微修正

不要なクラスとインターフェイスを削除

使われていない機能を別ファイルに移動

不要な引数を削除

警告に対処

同上

微修正

微修正

微修正

微修正

描画のエントリーポイントを集約中

使われていないイベントを削除

微修正

使われていないインターフェイスを削除

不要なインターフェイスを削除

不要なクラスを削除中

不要なクラスを削除

使ってないenumを削除

プロパティ化

* 道路ネットワーク編集がPLATEUARnStructureModelに依存しないようにした

* RoadNetworkEditingSystemをシングルトンではなくした

* EditingSystemを単一責任に分割中

同上

Object型をちゃんとした型に

同上

同上

同上

同上

同上

同上

同上

EditingSystem.EditSceneViewGuiを依存性注入に変更

同上

EditingSystem.EditTargetSelectButtonを依存性注入に変更

* fix : RnLineString.GetAdvancedPointFromBackがうまく動かないバグ修正. それに合わせて処理を共通化 & コメント修正

* EditingSystemを単一責任に分割

同上

同上

* リファクタ: 不要な部分を削除、コードの移動と整理、EditorDataへの依存を軽減

リファクタ中

不要な部分を削除中

同上

同上

NodeEditorDataが不要なので削除した

ptEditorDataが不要なので削除

微修正

EnterableEditorDataを移動

ExitablePointを移動

EditorDataへの依存を軽減中

交差点編集の機能を分離

無駄な部分を削除

リファクタ

* 交差点編集のトラックの矢印を表示する

交差点の線を描画中

同上

矢印を調整

交差点の選択を解除したら線が消えるように

* トラック生成機能を別クラスに分離した

ローカル関数をメンバー関数に

ローカル関数を外に出し中

同上

同上

引数整理中

引数を調整

同上

同上

途中

* 道路編集でトラックを作成したとき、単に直線で繋ぐのではなく、道路ネットワークのトラック生成の処理を使ってカーブもできるようにする

* マージ

* マージ

---------
