+++ Editor/Window/Main/Tab/RoadGUIParts/RoadEditPanel.cs
@@ -134,8 +134,10 @@ namespace PLATEAU.Editor.Window.Main.Tab.RoadGuiParts
         {
             
             rootVisualElement.Unbind();
+
             EditingSystem.EditTargetSelectButton.EnableLimitSceneViewDefaultContorl = false;
             EditingSystem.roadNetworkEditTarget.OnChangedSelectRoadNetworkElement -= OnChangedSelectedRoadBase;
+            
             
             RoadNetworkEditingSystem.TryTerminate(EditingSystem, rootVisualElement);
             if(editModeToggle != null) editModeToggle.value = false;
@@ -146,6 +148,8 @@ namespace PLATEAU.Editor.Window.Main.Tab.RoadGuiParts
         /// </summary>
         private void OnChangedSelectedRoadBase()
         {
+            EditingSystem.intersectionEditSceneViewGui.Terminate();
+            
             var roadGroupEditorData = EditingSystem.roadNetworkEditTarget.SelectedRoadNetworkElement as EditorData<RnRoadGroup>;
             if (roadGroupEditorData != null)
             {
@@ -162,7 +166,6 @@ namespace PLATEAU.Editor.Window.Main.Tab.RoadGuiParts
                 rootVisualElement.TrackSerializedObjectValue(selectedRoad, (se) =>
                 {
                     var mod = EditingSystem.roadEditSceneViewGui;
-
                     var obj = se as IScriptableRoadMdl;
                     if (mod.CanSetDtailMode())
                     {
@@ -213,7 +216,6 @@ namespace PLATEAU.Editor.Window.Main.Tab.RoadGuiParts
         private void OnApplyRoadButtonClicked()
         {
             bool isChanged = selectedRoad.Apply(EditingSystem.roadEditSceneViewGui);
-
             isChanged |= prevPlaceCrosswalkToggle != placeCrosswalkToggle.value;
             if (!isChanged)
             {
@@ -249,7 +251,6 @@ namespace PLATEAU.Editor.Window.Main.Tab.RoadGuiParts
             UpdateSplineButtonVisual(true);
             selectedRoad.IsSplineEditMode = true;
             selectedRoad.ApplySplineEditMode(EditingSystem?.roadEditSceneViewGui);
-
         }
 
         private void OnRoadSplineStopButtonClicked()
