+++ Runtime/RoadNetwork/Drawer/LaneLineDrawer.cs
@@ -3,10 +3,13 @@ using PLATEAU.Util;
 using PLATEAU.Util.GeoGraph;
 using System.Collections.Generic;
 using System.Linq;
+using UnityEditor;
 using UnityEngine;
 
 namespace PLATEAU.Editor.RoadNetwork.EditingSystemSubMod
 {
+    
+    
     /// <summary>
     /// 道路のレーン編集（交差点編集時の線も含む）のために、線1つをシーン上に描画します。
     /// 線の種類をサブクラスで書き分けます。
@@ -48,44 +51,71 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystemSubMod
     /// <summary>
     /// <see cref="ILaneLineDrawer"/>の矢印がリピートする版です。
     /// </summary>
-    public class LaneLineDrawerArrow : ILaneLineDrawer
+    internal class LaneLineDrawerArrow : ILaneLineDrawer
     {
         private RnWay way;
         private Color color;
-        public LaneLineDrawerArrow(RnWay way, Color color)
+        private LaneLineDrawMethod method;
+        private float oneSpaceLength;
+        public LaneLineDrawerArrow(RnWay way, Color color, LaneLineDrawMethod method, float oneSpaceLength = 1f)
         {
             this.way = way;
             this.color = color;
+            this.method = method;
+            this.oneSpaceLength = oneSpaceLength;
         }
 
         public void Draw()
         {
 #if UNITY_EDITOR
             const float OneArrowLength = 3f;
-            const float OneSpaceLength = 1f;
-            var prevColor = Gizmos.color;
-            Gizmos.color = color;
-            DrawDashedArrows(way.Select(v => v.PutY(v.y + ILaneLineDrawer.HeightOffset)), false, OneArrowLength, OneSpaceLength);
-            Gizmos.color = prevColor;
+            var prevColor = method switch
+            {
+                LaneLineDrawMethod.Gizmos => Gizmos.color,
+                LaneLineDrawMethod.Handles => Handles.color
+            };
+            
+            switch (method)
+            {
+                case LaneLineDrawMethod.Gizmos:
+                    Gizmos.color = color;
+                    break;
+                case LaneLineDrawMethod.Handles:
+                    Handles.color = color;
+                    break;
+            }
+            DrawDashedArrows(way.Select(v => v.PutY(v.y + ILaneLineDrawer.HeightOffset)), method, false, OneArrowLength, oneSpaceLength);
+            switch (method)
+            {
+                case LaneLineDrawMethod.Gizmos:
+                    Gizmos.color = prevColor;
+                    break;
+                case LaneLineDrawMethod.Handles:
+                    Handles.color = prevColor;
+                    break;
+            }
+
+
 #endif
         }
 
-        private static void DrawDashedArrows(IEnumerable<Vector3> vertices, bool isLoop = false, float lineLength = 3f, float spaceLength = 1f)
+        private static void DrawDashedArrows(IEnumerable<Vector3> vertices, LaneLineDrawMethod method, bool isLoop = false, float lineLength = 3f, float spaceLength = 1f)
         {
             foreach (var e in GeoGraphEx.GetEdges(vertices, isLoop))
-                DrawDashedArrow(e.Item1, e.Item2, lineLength, spaceLength);
+                DrawDashedArrow(e.Item1, e.Item2, method, lineLength, spaceLength);
         }
 
 
         private static void DrawArrow(
               Vector3 start
             , Vector3 end
+            , LaneLineDrawMethod method
             , float arrowSize = 0.5f
             , Vector3? arrowUp = null)
         {
             Vector3 up = arrowUp ?? Vector3.up;
 
-            DrawLine(start, end);
+            DrawLine(start, end, method);
             if (arrowSize > 0f)
             {
                 up = Vector3.Cross(end - start, Vector3.Cross(end - start, up)).normalized;
@@ -93,8 +123,8 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystemSubMod
                 var a2 = Quaternion.AngleAxis(-90, up) * a1;
                 a1 = a1.normalized;
                 a2 = a2.normalized;
-                DrawLine(end + a1 * arrowSize, end);
-                DrawLine(end + a2 * arrowSize, end);
+                DrawLine(end + a1 * arrowSize, end, method);
+                DrawLine(end + a2 * arrowSize, end, method);
             }
         }
         
@@ -102,13 +132,21 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystemSubMod
         /// <summary>
         /// Debug.DrawLineのラッパー. デバッグ描画系をここに集約するため
         /// </summary>
-        private static void DrawLine(Vector3 start, Vector3 end)
+        private static void DrawLine(Vector3 start, Vector3 end, LaneLineDrawMethod method)
         {
-            Gizmos.DrawLine(start, end);
+            switch (method)
+            {
+                case LaneLineDrawMethod.Gizmos:
+                    Gizmos.DrawLine(start, end);
+                    break;
+                case LaneLineDrawMethod.Handles:
+                    Handles.DrawLine(start, end);
+                    break;
+            }
         }
 
 
-        private static void DrawDashedArrow(Vector3 st, Vector3 en, float lineLength = 1f, float spaceLength = 0.2f, float arrowSize = 0.5f, Vector3? arrowUp = null)
+        private static void DrawDashedArrow(Vector3 st, Vector3 en, LaneLineDrawMethod method, float lineLength = 1f, float spaceLength = 0.2f, float arrowSize = 0.5f, Vector3? arrowUp = null)
         {
             var len = (en - st).magnitude;
 
@@ -123,11 +161,14 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystemSubMod
             {
                 var p0 = Vector3.Lerp(st, en, t);
                 var p1 = Vector3.Lerp(st, en, Mathf.Min(1f, t + s));
-                DrawArrow(p0, p1, arrowSize, arrowUp);
+                DrawArrow(p0, p1, method, arrowSize, arrowUp);
             }
         }
-        
+    }
 
+    internal enum LaneLineDrawMethod
+    {
+        Gizmos, Handles
     }
     
 }
